Regular adjacency matrix
This functions creates an adjacency matrix of "n" rows and "n" columns.

Each row/column is a node in the network. The intersection between a column a row represents their relationship. In this case, if it's 0, they are not connected whereas if is 1 they are connected to each other.
The code allows one to connected a node with its 4 nearest neighbors or with its 8 nearest neighbors (including diagonals).
It also allow to create a torus colum-wise.